+++ webapp/src/types/event-source-polyfill.d.ts
@@ -1,5 +1,0 @@
-declare module "event-source-polyfill" {
-  export class EventSourcePolyfill extends EventSource {
-    constructor(url: string, eventSourceInitDict?: EventSourceInit);
-  }
-}